chore: move functions module to `lib`
